refactor(config-ui): combine dir pages and routes
