+++ config-ui/src/routes/blueprint/connection-detail/index.tsx
