+++ config-ui/src/routes/blueprint/detail/blueprint-detail.tsx
