+++ config-ui/src/routes/blueprint/detail/components/advanced-editor/example/starrocks.ts
