+++ config-ui/src/routes/blueprint/detail/components/index.ts
