+++ config-ui/src/routes/blueprint/detail/components/sync-policy/index.tsx
