+++ config-ui/src/routes/blueprint/detail/components/sync-policy/styled.ts
