+++ config-ui/src/routes/blueprint/detail/configuration-panel.tsx
