+++ config-ui/src/routes/blueprint/detail/index.ts
