+++ config-ui/src/routes/blueprint/detail/status-panel.tsx
