+++ config-ui/src/routes/blueprint/detail/styled.ts
