+++ config-ui/src/routes/index.ts
@@ -16,5 +16,10 @@
  *
  */
 
+export * from './api-keys';
 export * from './blueprint';
+export * from './connection';
+export * from './error';
+export * from './layout';
+export * from './pipeline';
 export * from './project';
