+++ config-ui/src/routes/project/detail/settings-panel.tsx
