+++ config-ui/src/routes/project/detail/styled.ts
