+++ config-ui/src/routes/project/detail/webhooks-panel.tsx
